Added Spider/Radar Chart Plot
* Added show_spider_chart(), which takes a dictionary containing labels (str) as keys and floats as values and plots a radar/spider chart.
This form of plot will be used to display the different ELO scores that a user has for the different categories of problems that they have tried.
Whenever they solve a problem belonging to a certain category, only the ELO of that category will be affected, leaving the rest of ELOs displayed in the chart untouched.

* Added show_bar_plot(), which takes two 1D arrays (and other optional args) and plots a bar graph.